got it working but root doenst match and there is not enough gas to go above depth 4

+++ contracts/evm/L1WarpToad.sol
@@ -6,7 +6,7 @@ contract L1WarpToad is WarpToadCore {
     address nativeToken;
 
     //@joss do what ever to the structure or naming or things if u feel like it
-    constructor(uint256 _maxTreeDepth, address _gigaBridge,address _nativeToken, string memory name, string memory symbol) 
+    constructor(uint8 _maxTreeDepth, address _gigaBridge,address _nativeToken, string memory name, string memory symbol) 
     ERC20(name, symbol) 
     WarpToadCore(_maxTreeDepth, _gigaBridge)  {
         nativeToken = _nativeToken;
