+++ contracts/evm/L2WarpToad.sol
@@ -4,7 +4,7 @@ import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract L2WarpToad is WarpToadCore {
     //@joss do what ever to the structure or naming or things if u feel like it
     constructor(
-        uint256 _maxTreeDepth,
+        uint8 _maxTreeDepth,
         address _gigaBridge,
         string memory name,
         string memory symbol
